(polybar) Replace spotify scrolling status with fixed
The scrolling was a little bit resource heavy for what it provided. It
also happened to be such that the `playerctl status` command exhausted
an API-key for `spotifyd` which the spotify-tui command `spt playback
-s` doesn't do. However, it's a little bit annoying with the title
dissapearing after being paused for a while and also not being able to
reume with media keys in those events.